requirements.txt
lime==0.2.0.1
lightgbm==3.1.1
shap==0.37.0
numpy==1.19.5
Flask==1.1.2
altair==4.1.0
pandas==1.2.1
matplotlib==3.3.2
pydeck==0.5.0
requests==2.25.1
seaborn==0.11.1
xgboost==1.3.1
streamlit==0.74.1
Pillow==8.1.0
scikit_learn==0.24.1
